scripts/requirements.txt
openai
airtable-python-wrapper
python-dotenv
